+++ code/src/cljs/sixsq/nuvla/ui/deployments/views.cljs
@@ -46,10 +46,11 @@
         ^{:key (random-uuid)}
         [:div {:style {:margin-top "10px"}}
          [filter-comp/ButtonFilter
-          {:resource-name  "deployment"
+          {:resource-name  spec/resource-name
            :default-filter @additional-filter
            :open?          filter-open?
-           :on-done        #(dispatch [::events/set-additional-filter %])}]]]])))
+           :on-done        #(dispatch [::events/set-additional-filter %])
+           :show-clear-button-outside-modal? true}]]]])))
 
 (defn BulkUpdateModal
   []
@@ -135,7 +136,6 @@
                         :icon     "ellipsis vertical"
                         :disabled (not (pos? @selected-count))}
            [ui/DropdownMenu
-            #_[ui/DropdownItem "Start"]
             [ui/DropdownItem
              {:on-click #(dispatch [::events/bulk-update-params])} (str/capitalize (@tr [:update]))]
             ^{:key @modal-stop-key}