+++ CHANGELOG.md
@@ -2,12 +2,28 @@
 
 ## Unreleased
 
+- Notificaiton subscriptions - Shows GiB unit for network in modal
+- Api Page - Persists and loads all filter parameters
 - Tabs - New design
 - Apps Store, Navigate apps tab - New design
 - Apps Project Details, Overview tab - New design
 - Apps Details, Details tab - New design
+- Apps Details, Share tab - New design
 
-## [2.33.2] - 2023-02-24
+## [2.33.3] - 2023-02-27
+
+### Changed
+
+- Deployment modal - Display infrastructure docker labels #996
+- Deployment modal - Do not allow deployment of Swarm application on Swarm disabled node or worker node #996
+- Deployment modal - Inform user that he is deploying a Compose application on a Docker swarm node #996
+- Clouds and cloud details - Show compatibility label
+- Logger - Add e2e test
+- Sign in/up - Removes border from buttons, bugfix #1177
+- Filter wizard - Additional filter is stored in and loads from query param
+- Filter wizard - Reopening modal with active filter shows correct attribute and condition in dropdown, bugfix #1159
+
+## [2.33.2] - 2023-02-27
 
 ### Changed